% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R, R/pcir.package.R
\name{bubble}
\alias{bubble}
\title{Create a Bubble Plot for PCI Visualization}
\usage{
bubble(df3)

bubble(df3)
}
\arguments{
\item{df3}{A data frame generated by the \code{pci} function.}
}
\value{
A ggplot2 object representing the bubble plot.

A ggplot2 object representing the bubble plot.
}
\description{
This function generates a bubble plot to visualize the results of the PCI
calculation. It shows the mean action acceptability on the y-axis and the PCI
value as the size of the bubbles.

Create a bubble plot to visualize PCI results.
}
\examples{
df3 <- pci(df2)
p <- bubble(df3)
print(p)
df3 <- pci(df2)
bubble(df3)
}
