% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcir.package.R
\docType{package}
\name{pcir-package}
\alias{pcir}
\alias{pcir-package}
\title{pcir Package Functions and Setup}
\description{
This file contains the basic setup and utility functions for the 'pcir' package.
The pcir package is designed for calculating, comparing, and visualizing the
Potential for Conflict Index (PCI).
}
\section{Functions}{

\itemize{
\item \code{counting()}: Summarize data by calculating counts, percentages, means, and standard deviations.
\item \code{pci()}: Compute the Potential for Conflict Index (PCI) from summary data.
\item \code{bubble()}: Create a bubble plot to visualize PCI results.
}
}

\author{
\strong{Maintainer}: Francesca Palmeira \email{francesca@alumni.usp.br} (\href{https://orcid.org/0000-0002-7597-1157}{ORCID})

}
