% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pci.R
\name{pci}
\alias{pci}
\title{Calculate the Potential for Conflict Index (PCI)}
\usage{
pci(df2)
}
\arguments{
\item{df2}{A data frame generated by the \code{counting} function, containing the counts and related statistics.}
}
\value{
A data frame with the calculated PCI values for each group.
}
\description{
This function calculates the Potential for Conflict Index (PCI) based on the counts provided. PCI is useful in understanding the level of consensus and conflict in the data.
}
\examples{
df2 <- counting(df1)
pci_values <- pci(df2)
print(pci_values)
}
