% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pci.R, R/pcir.package.R
\name{pci}
\alias{pci}
\title{PCI Function}
\usage{
pci(df2)

pci(df2)
}
\arguments{
\item{df2}{A data frame generated by the \code{counting} function.}
}
\value{
A data frame with the calculated PCI values for each group.

A data frame with the calculated PCI values for each group.
}
\description{
Calculate the Potential for Conflict Index (PCI).

Calculate the Potential for Conflict Index (PCI).
}
\examples{
df2 <- counting(df1)
pci(df2)
df2 <- counting(df1)
pci(df2)
}
