% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcir.package.R
\docType{package}
\name{pcir-package}
\alias{pcir}
\alias{pcir-package}
\title{pcir Package Functions and Setup}
\description{
This file contains the basic setup and utility functions for the 'pcir' package.
The pcir package is designed for calculating, comparing, and visualizing the
Potential for Conflict Index (PCI).
}
\section{Functions}{

- `counting()`: Summarize data by calculating counts, percentages, means, and standard deviations.
- `pci()`: Compute the Potential for Conflict Index (PCI) from summary data.
- `bubble()`: Create a bubble plot to visualize PCI results.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fblpalmeira/pcir}
  \item Report bugs at \url{https://github.com/fblpalmeira/pcir/issues}
}

}
